APIv4 - Add EntitySet api for set-ops such as UNION
