+++ Civi/Api4/Query/SqlEquation.php
@@ -76,10 +76,10 @@ class SqlEquation extends SqlExpression {
   /**
    * Render the expression for insertion into the sql query
    *
-   * @param \Civi\Api4\Query\Api4SelectQuery $query
+   * @param \Civi\Api4\Query\Api4Query $query
    * @return string
    */
-  public function render(Api4SelectQuery $query): string {
+  public function render(Api4Query $query): string {
     $output = [];
     foreach ($this->args as $i => $arg) {
       // Just an operator
