+++ Civi/Api4/Query/SqlString.php
@@ -27,7 +27,7 @@ class SqlString extends SqlExpression {
     $this->expr = str_replace(['\\\\', "\\$quot", $backslash], [$backslash, $quot, '\\\\'], $str);
   }
 
-  public function render(Api4SelectQuery $query): string {
+  public function render(Api4Query $query): string {
     return '"' . \CRM_Core_DAO::escapeString($this->expr) . '"';
   }
