+++ Civi/Api4/Query/SqlWild.php
@@ -19,7 +19,7 @@ class SqlWild extends SqlExpression {
   protected function initialize() {
   }
 
-  public function render(Api4SelectQuery $query): string {
+  public function render(Api4Query $query): string {
     return '*';
   }
